[chore][한정빈] : (main)/layout.tsx 추가
