+++ app/(main)/layout.tsx
@@ -1,33 +1,3 @@
-import type {Metadata} from 'next';
-import {Geist, Geist_Mono} from 'next/font/google';
-import localFont from 'next/font/local';
-import './globals.css';
-
-const geistSans = Geist({
-  variable: '--font-geist-sans',
-  subsets: ['latin'],
-});
-
-const geistMono = Geist_Mono({
-  variable: '--font-geist-mono',
-  subsets: ['latin'],
-});
-
-const pretendard = localFont({
-  src: '../fonts/PretendardVariable.woff2',
-  display: 'swap',
-  variable: '--font-pretendard',
-});
-
-export const metadata: Metadata = {
-  title: 'Create Next App',
-  description: 'Generated by create next app',
-};
-
-export default function RootLayout({children}: Readonly<{children: React.ReactNode}>) {
-  return (
-    <html lang="en">
-      <body className={`${geistSans.variable} ${geistMono.variable} ${pretendard.className} antialiased`}>{children}</body>
-    </html>
-  );
+export default function MainLayout({children}: {children: React.ReactNode}) {
+  return <div>{children}</div>;
 }
